[Refactor Artist] Move Artist related components to /Apps/Artist
