+++ src/Apps/Artist/Components/ArtistCollectionsRail/ArtistCollectionEntity.tsx
