+++ src/Apps/Artist/Components/ArtistCollectionsRail/ArtistCollectionsRail.tsx
