+++ src/Apps/Artist/Components/MarketDataSummary/index.tsx
