+++ src/Apps/Artist/Components/MarketInsights/__tests__/Fixtures/Artists.fixture.ts
@@ -1,5 +1,5 @@
-import { MarketDataSummary_artist } from "../../../__generated__/MarketDataSummary_artist.graphql"
-import { MarketInsights_artist } from "../../../__generated__/MarketInsights_artist.graphql"
+import { MarketDataSummary_artist } from "__generated__/MarketDataSummary_artist.graphql"
+import { MarketInsights_artist } from "__generated__/MarketInsights_artist.graphql"
 
 export const MarketDataSummaryArtists: MarketDataSummary_artist[] = [
   {
