+++ src/Apps/Artist/Components/MarketInsights/index.tsx
