+++ src/Apps/Artist/__stories__/MarketInsights.story.tsx
@@ -1,8 +1,8 @@
 import { artistResponse } from "Apps/__tests__/Fixtures/MarketInsights"
+import { MarketInsights } from "Components/v2/ArtistMarketInsights"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Utils/Section"
-import { MarketInsights } from "../ArtistMarketInsights"
 
 storiesOf("Styleguide/Components", module).add("MarketInsights", () => {
   return (
